Extract out SourceLocation into own class
