+++ src/main/java/teammates/common/datatransfer/GeneralLogEntry.java
@@ -1,9 +1,10 @@
 package teammates.common.datatransfer;
 
 import java.util.Map;
-import java.util.Objects;
 
 import javax.annotation.Nullable;
+
+import teammates.common.datatransfer.logs.SourceLocation;
 
 /**
  * Represents a log entry and contains the fields that are more important
@@ -77,49 +78,4 @@ public class GeneralLogEntry {
         return details;
     }
 
-    /**
-     * Represents a location of source code that produces a log line.
-     */
-    public static class SourceLocation {
-        private final String file;
-        private final Long line;
-        private final String function;
-
-        public SourceLocation(String file, Long line, String function) {
-            this.file = file;
-            this.line = line;
-            this.function = function;
-        }
-
-        public String getFile() {
-            return file;
-        }
-
-        public Long getLine() {
-            return line;
-        }
-
-        public String getFunction() {
-            return function;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (obj == this) {
-                return true;
-            }
-            if (obj instanceof SourceLocation) {
-                SourceLocation other = (SourceLocation) obj;
-                return file.equals(other.getFile())
-                        && line.equals(other.getLine())
-                        && function.equals(other.getFunction());
-            }
-            return false;
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(file, line, function);
-        }
-    }
 }
