+++ src/main/java/teammates/ui/webapi/QueryLogsAction.java
@@ -9,11 +9,11 @@ import java.util.stream.Collectors;
 import org.apache.http.HttpStatus;
 
 import teammates.common.datatransfer.GeneralLogEntry;
-import teammates.common.datatransfer.GeneralLogEntry.SourceLocation;
 import teammates.common.datatransfer.QueryLogsParams;
 import teammates.common.datatransfer.QueryLogsParams.UserInfoParams;
 import teammates.common.datatransfer.QueryLogsResults;
 import teammates.common.datatransfer.logs.LogEvent;
+import teammates.common.datatransfer.logs.SourceLocation;
 import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.exception.LogServiceException;
 import teammates.common.exception.UnauthorizedAccessException;
